requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
gunicorn==22.0.0
azure-storage-blob==12.23.1
azure-cosmos==4.7.0
requests==2.32.3
